restructured types
