+++ src/adminDashboard/services.ts
@@ -170,6 +170,89 @@ export interface RevenuePoolAccount {
       }
   };
 
+    /**
+   * Initialize a prize pool for a specific tournament
+   * @param tournamentId - The tournament ID
+   * @param mintPublicKey - The token mint address
+   * @returns Result object with transaction details and addresses
+   */
+    export const initializePrizePoolService = async (tournamentId: string, mintPublicKey: PublicKey, adminPublicKey: PublicKey) => {
+        try {
+          const { program, connection } = getProgram();
+      
+          // Log initial parameters for clarity
+          console.log("Initializing Prize Pool for Tournament:");
+          console.log("Tournament ID:", tournamentId);
+          console.log("Admin PublicKey:", adminPublicKey.toBase58());
+          console.log("Mint PublicKey:", mintPublicKey.toBase58());
+      
+          // First, derive the tournament pool PDA to ensure it exists
+          const tournamentIdBytes = Buffer.from(tournamentId, "utf8");
+          const [tournamentPoolPublicKey] = PublicKey.findProgramAddressSync(
+            [Buffer.from("tournament_pool"), adminPublicKey.toBuffer(), tournamentIdBytes],
+            program.programId
+          );
+          
+          console.log("🔹 Tournament Pool PDA Address:", tournamentPoolPublicKey.toString());
+          
+          // Add this to initializePrizePoolService
+          console.log("Full tournament pool key:", tournamentPoolPublicKey.toString());
+          console.log("Tournament ID bytes:", tournamentIdBytes);
+          console.log("Admin pubkey:", adminPublicKey.toString());
+      
+          // Derive the PDA for the prize pool (now derived from tournament pool)
+          const [prizePoolPublicKey] = PublicKey.findProgramAddressSync(
+            [Buffer.from("prize_pool"), tournamentPoolPublicKey.toBuffer()],
+            program.programId
+          );
+      
+          // Derive the PDA for the prize escrow account
+          const [prizeEscrowPublicKey] = PublicKey.findProgramAddressSync(
+            [Buffer.from("prize_escrow"), prizePoolPublicKey.toBuffer()],
+            program.programId
+          );
+      
+          console.log("🔹 Prize Pool PDA Address:", prizePoolPublicKey.toString());
+          console.log("🔹 Prize Escrow PDA Address:", prizeEscrowPublicKey.toString());
+      
+          // Get the latest blockhash
+          const { blockhash } = await connection.getLatestBlockhash("finalized");
+          console.log("Latest Blockhash:", blockhash);
+      
+          // Create the transaction
+          const transaction = await program.methods
+            .initializePrizePool(tournamentId)
+            .accounts({
+              prizePool: prizePoolPublicKey,
+              tournamentPool: tournamentPoolPublicKey,
+              prizeEscrowAccount: prizeEscrowPublicKey,
+              mint: mintPublicKey,
+              admin: adminPublicKey,
+              systemProgram: anchor.web3.SystemProgram.programId,
+              tokenProgram: TOKEN_2022_PROGRAM_ID,
+            })
+            .transaction();
+      
+          // Set recent blockhash and fee payer
+          transaction.recentBlockhash = blockhash;
+          transaction.feePayer = adminPublicKey;
+          // Serialize transaction and send it to the frontend
+        return {
+          success: true,
+          message: "Transaction created successfully!",
+          transaction: transaction.serialize({ requireAllSignatures: false }).toString('base64'),
+          prizePool: prizePoolPublicKey.toString(),
+        };
+      } catch (err) {
+          console.error("❌ Error initializing prize pool:", err);
+          return {
+            success: false,
+            message: `Error initializing prize pool: ${err.message || err}`
+          };
+        }
+      };
+      
+
 
   // ✅ Initialize Reward Pool (admin-only)
 export const initializeRewardPoolService = async (
