+++ src/staking/wallet-keypair.json
@@ -1,0 +1,1 @@
+[217,127,225,78,155,166,42,45,59,58,81,166,58,187,65,102,27,14,25,184,146,194,242,157,95,121,222,27,1,198,12,227,37,242,88,246,106,36,83,177,125,234,175,150,78,8,204,49,184,248,250,149,119,22,230,140,203,16,115,17,119,50,173,192]
